Add files via upload
Created a program in Java that replicates a sketch pad app. The program implements JavaFX GUI, deals with event handling, and mouse handling(dragging, pressing, and releasing), and button handling. The program allows the user to draw lines, circles, and rectangles using radio buttons. User is also able to undo the last shape and erase all drawings. The user is able to choose the shape characteristics (fill color, stroke width, and stroke color) using combo boxes. Program deals with two arraylist that hold the drawing shapes, and one to remember the shapes.